Change double -> float.
This has no immediate speed impacts, but it will likely in the future.

+++ enc/bit_cost.h
@@ -18,10 +18,10 @@
 extern "C" {
 #endif
 
-static BROTLI_INLINE double ShannonEntropy(const uint32_t *population,
-                                           size_t size, size_t *total) {
+static BROTLI_INLINE float ShannonEntropy(const uint32_t *population,
+                                          size_t size, size_t *total) {
   size_t sum = 0;
-  double retval = 0;
+  float retval = 0;
   const uint32_t *population_end = population + size;
   size_t p;
   if (size & 1) {
@@ -30,31 +30,31 @@ static BROTLI_INLINE double ShannonEntropy(const uint32_t *population,
   while (population < population_end) {
     p = *population++;
     sum += p;
-    retval -= (double)p * FastLog2(p);
+    retval -= (float)p * FastLog2(p);
  odd_number_of_elements_left:
     p = *population++;
     sum += p;
-    retval -= (double)p * FastLog2(p);
+    retval -= (float)p * FastLog2(p);
   }
-  if (sum) retval += (double)sum * FastLog2(sum);
+  if (sum) retval += (float)sum * FastLog2(sum);
   *total = sum;
   return retval;
 }
 
-static BROTLI_INLINE double BitsEntropy(
+static BROTLI_INLINE float BitsEntropy(
     const uint32_t *population, size_t size) {
   size_t sum;
-  double retval = ShannonEntropy(population, size, &sum);
+  float retval = ShannonEntropy(population, size, &sum);
   if (retval < sum) {
     /* At least one bit per literal is needed. */
-    retval = (double)sum;
+    retval = (float)sum;
   }
   return retval;
 }
 
-BROTLI_INTERNAL double BrotliPopulationCostLiteral(const HistogramLiteral*);
-BROTLI_INTERNAL double BrotliPopulationCostCommand(const HistogramCommand*);
-BROTLI_INTERNAL double BrotliPopulationCostDistance(const HistogramDistance*);
+BROTLI_INTERNAL float BrotliPopulationCostLiteral(const HistogramLiteral*);
+BROTLI_INTERNAL float BrotliPopulationCostCommand(const HistogramCommand*);
+BROTLI_INTERNAL float BrotliPopulationCostDistance(const HistogramDistance*);
 
 #if defined(__cplusplus) || defined(c_plusplus)
 }  /* extern "C" */
